PieceManager.h: getRandomPiece() added, pieceIsReadyToDownload() to check conditions for download; sendMessage() for sending a choke message directly after Hadnshake added in processHadnshake(); added some recursive logic into write() in tcp.h - has to be adjusted to the general programm and boost::asio logic

+++ include/torrent/protocol/Piece.h
@@ -47,7 +47,7 @@ namespace torrent::protocol{
   };
   class Piece{
   public:
-    int32_t getIndex();
+    size_t getIndex();
     size_t getPieceSize();
     bool isCompleted();
     bool hasBlockToDownload();
